pyproject.toml
[project]
name = "civic-cli-tools"
version = "0.6.0"
description = "Policy research CLI — evidence-based briefs from 8 government sources"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0",
    "exa-py>=1.0",
    "httpx>=0.27",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.scripts]
civic = "cli:main"

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
sources = ["src"]

[tool.hatch.build.targets.wheel.force-include]
"topics.toml" = "topics.toml"
"LICENSE" = "LICENSE"
